Extracts SubscriberStub to testkit-core.
